+++ CHANGELOG.md
@@ -14,6 +14,10 @@
 
 ### Updated
 
+## [1.0.0-alpha.9] - 2022-06-14
+
+### Updated
+
 - `com.javiersc.kotlin:kotlin-stdlib -> 0.1.0-alpha.5`
 
 ## [1.0.0-alpha.8] - 2022-06-14